Move fixSelectionInZeroWidthBlock
