+++ packages/slate-react/src/plugins/android/index.js
@@ -1,8 +1,43 @@
 import getWindow from 'get-window'
-import fixSelectionInZeroWidthBlock from './fix-selection-in-zero-width-block'
 import CompositionManager from './composition-manager'
 
-function MutationPlugin({ editor }) {
+/**
+ * Fixes a selection within the DOM when the cursor is in Slate's special
+ * zero-width block. Slate handles empty blocks in a special manner and the
+ * cursor can end up either before or after the non-breaking space. This
+ * causes different behavior in Android and so we make sure the seleciton is
+ * always before the zero-width space.
+ *
+ * @param {Window} window
+ */
+
+function fixSelectionInZeroWidthBlock(window) {
+  const domSelection = window.getSelection()
+  const { anchorNode } = domSelection
+  if (anchorNode == null) return
+  const { dataset } = anchorNode.parentElement
+  const isZeroWidth = dataset ? dataset.slateZeroWidth === 'n' : false
+
+  if (
+    isZeroWidth &&
+    anchorNode.textContent.length === 1 &&
+    domSelection.anchorOffset !== 0
+  ) {
+    const range = window.document.createRange()
+    range.setStart(anchorNode, 0)
+    range.setEnd(anchorNode, 0)
+    domSelection.removeAllRanges()
+    domSelection.addRange(range)
+  }
+}
+
+/**
+ * Android Plugin
+ *
+ * @param {Editor} options.editor
+ */
+
+function AndroidPlugin({ editor }) {
   const observer = new CompositionManager(editor)
 
   function onCompositionStart() {
@@ -50,4 +85,4 @@ function MutationPlugin({ editor }) {
   }
 }
 
-export default MutationPlugin
+export default AndroidPlugin
